backend/requirements.txt
#Это файл зависимостей.
fastapi==0.115.0 #Веб фреймворк для создания API на python. Обрабатывает HTTP-запросы (GET,POST,PUT,DELETE)
uvicorn[standard]==0.30.1 # ASGI-сервер. Слушает входящие сетевые запросы и передает их FastAPI.
sqlalchemy==2.0.35 #ORM Позволяет работать с БД PostgreSQL. Мы делваем не сырые запросы к БД, а описываем таблицы как python классы, а она уже генерирует и выполняет SQL код.
psycopg2-binary==2.9.11 #Адаптер для подключения python к БД. Помогает alchemy разговаривать с БД на одном языке.
python-dotenv==1.0.1 #Библиотека для загрузки переменных окружения из файла .env 